Garbage collection now occrus after an action finishes to free up system resources; Small fix made to how entity dtype is loaded in; Changed some unitialised block data to np.inf from np.nan to avoid conflicting with numerical sim output; Process are now forcefully terminated when they report that they have finished.

+++ utils/entity.py
@@ -40,8 +40,9 @@ class Entity:
     
     def CleanUp(self):
         # remove the data from memory to stop it persisting after closing the application.
-        self.dataSharedMemory.close()
-        self.dataSharedMemory.unlink()
+        if hasattr(self, 'dataSharedMemory'):
+            self.dataSharedMemory.close()
+            self.dataSharedMemory.unlink()
 
     def Register(self, overrideID = None):
         '''Registers this object as an entity inside the shared entity list.'''
